ZINQ for benchdamic

zinq adapts the zinq method for integration with the benchdamic package.

DA_ZINQ(
  object,
  pseudo_count = FALSE,
  conditions_col,
  conditions,
  norm = "none",
  pval_method,
  y_CorD,
  verbose = FALSE
)

Arguments

object

A (Tree)SummarizedExperiment or a phyloseq object.

pseudo_count

Whether include or not a pseudo_count. Default = FALSE.

conditions_col

The name of the grouping column, which is located in the colData (SummarizedExperiment) or the sample_data (phyloseq).

conditions

A character vector indicating the conditions. Example: c(condB = 'control', condA = 'treatment')

norm

A character string indicating the normalization method to be used prior analysis. Default is 'none'.

pval_method

A character string indicating the type of pvalue to use. Options: Cauchy or MinP. Default: Cauchy.

y_CorD

Indicate if data are counts 'D' or continuous 'C'.

verbose

Whether include messages or not. Default is FALSE.

Value

A list in the format used in the benhcdamic pipeline.

References

Ling, W., Zhao, N., Plantinga, A.M. et al. Powerful and robust non-parametric association testing for microbiome data via a zero-inflated quantile approach (ZINQ). Microbiome 9, 181 (2021). https://doi.org/10.1186/s40168-021-01129-3