Lefser method

DA_lefser is a modified version of the lefser package, which includes the pvalues of the Kruskal test.

DA_lefse(
  object,
  pseudo_count = FALSE,
  conditions,
  norm = "none",
  verbose = FALSE,
  groupCol,
  ...
)

Arguments

object

A phyloseq or (Tree)SummarizedExperiment object.

pseudo_count

Add or not a pseudocount of 1. Default is FALSE.

conditions

A named character vector of length two. Names must be "condB" and "condA" in that order. "condB" must indicate control/reference/denominator and "condA" must indicate treatment/target/numerator. Example: c(condB = 'control', condA = 'treatment')

norm

Normalization method. Options: none, CLR, and TSS.

verbose

If TRUE, messages are displayed on screen. Default is FALSE.

groupCol

Name of the column in colData with the conditions.

...

Parameters passed to lefser.

Value

An object ready to be included in the benchdamic framework.