Wilcox test for differential abundance

DA_wilcox performs Wilcoxon test on a phyloseq object.

Usage

DA_wilcox(
  object,
  pseudo_count = FALSE,
  norm = "none",
  conditions_col,
  conditions,
  verbose = FALSE
)

Arguments

object

A phyloseq object.

pseudo_count

Whether include a pseudocount or not. Default is FASLE.

norm

String character. Choose between three normalization methods: 'none', 'CLR', or 'TSS'.

conditions_col

String character. The name of the column in sample data with the conditions.

conditions

A named character vector. The names must be "condB" and "condA". condB indicates the reference/numerator/control and condA indicates the target/denominator/treatment. For example: c(condB = 'control', condA = 'treatment')

verbose

Print messages or not, Default is FALSE.

Value

A list with outputs compatible with the benchdamic framework.