CLR normalization

norm_clr applies a centered-log ratio (CLR) transformation to a matrix column-wise. Features (e.g. taxa, OTUs) must be in the rows and samples in the columns.

norm_clr(mat, pseudocount = 0, log = TRUE)

Arguments

mat

A count matrix.

pseudocount

Pseudocount added to the matrix. Default value is 0 (no pseudocount)..

log

If TRUE, CLR will be logged. Default = TRUE. In most cases this should be set to TRUE

Value

A CLR-transformed matrix.